python/plotly_dash/requirements.txt
numpy
pandas
dash_core_components
dash_html_components
pandas_datareader
dash_renderer
